refactor sso/settings.go into two files
All the SSOConfig related stuff is in sso/config.go now
